Intermediate step at writing class Population
Some code added to Relation as well. Two bugs found in class
Specie:

1. The global dictionary of parameters does not work (i.e., I
   failed to find any way to assign values to it).
2. The .reactions method is non-commutative ("2".reactions("a")
   gives a list different from "a".reactions("2"), reproducible
   using tests/testSpecie).

The first bug is put aside by commenting out all queries to the
dictionary and replacing them with real constants.

The second bug needs Lisa's hands.

Other than that, the new methods of class Population should not
be considered to be debugged until I check the whole output of
testTotalPopulation.

	modified:   population.cpp
	modified:   population.h
	modified:   reaction.cpp
	modified:   reaction.h
	modified:   relation.cpp
	modified:   relation.h
	modified:   specie.cpp
	modified:   tests/testSpecie.cpp
	modified:   totalPopulation.cpp