Better samplers at Population and Relation levels
Now if the remaining juice is greater than the sum of partial
propensities of the level (m_ksi for Population and m_psi for
Relation), the sampler returns an invalid reaction which causes
the top-level sampler (at TotalPopulation) to re-generate the
random number. This should decrease the errors associated with
the increase in the probability of the last subelement of the
current element (Relation for Populations, Reaction for Relations)
being selected.